requirements.txt
# MCP Framework
mcp>=1.0.0

# AI APIs
anthropic>=0.18.0
openai>=1.0.0

# Document Processing
PyPDF2>=2.0.0
python-docx>=0.8.11

# Vector Search and ML
numpy>=1.21.0,<2.0.0
scikit-learn>=1.0.0,<2.0.0

# Web Crawling and Scraping
requests>=2.28.0
beautifulsoup4>=4.11.0
urllib3>=1.26.0

# Environment variables
python-dotenv>=0.19.0
